# kb-rag-deploy/docs/openapi/kb-server.yaml
openapi: 3.0.3
info:
  title: kb-rag-server API
  description: 'kb-rag 管理台内部 API（M1 基线 + M2–M8 增量）。


    本文件按 docs/M1-CONTRACTS.md §5、docs/M2-CONTRACTS.md §1.5/§3/§4/§6 与

    docs/M3-CONTRACTS.md §3 定型的端点与 DTO 编写，用于 kb-rag-server / kb-rag-web

    之间的接口契约测试与前端 mock。

    M2 增量：search 新入参（改写/融合/重排/阈值/metadata_filter）与出参（applied、

    各路归一化分）、ik 词典 CRUD、知识库索引配置更新与按新配置重建。ik 词典热更新

    通道 `GET /internal/dict/ik/{ext|stop}.txt`（见 M2-CONTRACTS.md §3）面向 ES ik

    插件而非管理台前端，免鉴权、纯文本响应，不纳入本文件的管理台 API 契约。

    M3 增量：model-status 新增 vision_* 三字段；解析预览与确认（PENDING_CONFIRM 状态、

    单个/批量确认、改规则重解析）；聊天记录导入（匹配预览 + 确认执行）；告警 Webhook

    配置与测试；Demo 一键导入与状态查询；IndexConfig 新增 clean_rules/parse_preview_required/

    chat_aggregation。kb-parser 侧 VLM 调用与图片/清洗均在 kb-rag-server 完成，parser 不

    调用任何大模型（docs/M3-CONTRACTS.md §0/§1）。

    M4a 增量：文档版本列表与激活影响预检（`GET .../versions`、`GET .../activate-impact`、
    `POST .../activate`）、跨版本待复核标注清单、分片标注四操作（编辑/启禁用/合并/拆分）。

    M4b 增量（docs/M4b-CONTRACTS.md）：评测体系全链路——评测集与 case 的增删改查、
    证据复核工作台（`stale-cases`/`recheck`）、Demo 示例评测集导入（`import-demo`）、
    检索调试页一键收进评测集（`cases/from-retrieval`）与检索结果反馈标注
    （`retrieval-feedback`）、评测运行配置矩阵与费用预估（`runs`/`runs/estimate`）、
    run 详情/命中明细下钻/多 run 对比（`eval-runs/{runId}`、`.../results`、
    `eval-runs/compare`）；`ActivateImpact.affected_eval_case_count` 由恒 0 占位改为
    真实统计。对外 API Key 网关（应用发布渠道）属于 M4c 范畴，不在本文件中。
    M4c 增量（docs/M4c-CONTRACTS.md）：应用中心全链路——应用 CRUD（`/api/v1/apps`）、版本创建/
...
    输出与候选白名单求交集，空交集/解析失败/超时/未配置模型一律降级检索全部关联库并记
...

...

...
servers:
...
security:
- bearerAuth: []
paths:
...
  /api/v1/kb:
    post:
      tags:
      - knowledge-base
      summary: 创建知识库
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateKbRequest'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeBaseResult'
        '400':
          $ref: '#/components/responses/InvalidParam'
    get:
      tags:
      - knowledge-base
      summary: 知识库列表
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeBaseListResult'
  /api/v1/kb/{kbId}:
    parameters:
    - $ref: '#/components/parameters/KbId'
    get:
      tags:
      - knowledge-base
      summary: 知识库详情
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeBaseResult'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
      - knowledge-base
      summary: 删除知识库
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResult'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/kb/{kbId}/index-config:
    parameters:
    - $ref: '#/components/parameters/KbId'
    put:
      tags:
      - knowledge-base
      summary: 更新知识库索引配置（分片参数/父子分片）
      description: 'M2-CONTRACTS.md §4：更新后重算 current_config_fingerprint，并对该知识库下

        活跃版本 chunk_fingerprint 不等于新指纹的文档置 config_stale=1（web 侧据此

        展示“N 篇文档使用旧配置”提示 + 重建入口）。仅更新配置元数据，**不会**立即

        触发重建，重建需显式调用 `POST /api/v1/kb/{kbId}/rebuild`。

        '
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IndexConfig'
      responses:
        '200':
          description: 更新成功，返回最新知识库信息（含新指纹与索引配置）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeBaseResult'
        '400':
          $ref: '#/components/responses/InvalidParam'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/kb/{kbId}/rebuild:
    parameters:
    - $ref: '#/components/parameters/KbId'
    post:
      tags:
      - knowledge-base
      summary: 按当前索引配置重建文档索引
      description: 'M2-CONTRACTS.md §4：body 可选 `doc_ids`，缺省时重建该知识库下全部

        config_stale=1 的文档。每篇文档一个 REBUILD 任务：从 MinIO 解析产物

        重切/重嵌入，同一物理索引内原子替换该版本 chunk（先写新 chunk 再删旧，

        引擎双写走既有同步机制），完成后置 config_stale=0。异步受理，幂等。

        '
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RebuildRequest'
      responses:
        '200':
          description: 已受理
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResult'
        '400':
          $ref: '#/components/responses/InvalidParam'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/kb/{kbId}/documents/confirm:
    parameters:
    - $ref: '#/components/parameters/KbId'
    post:
      tags:
      - parse-preview
      summary: 批量确认解析预览
      description: 'M3-CONTRACTS.md §3.4：body 可选 `doc_ids`，缺省时确认该知识库下全部 PENDING_CONFIRM
        状态的文档。仅在知识库开启 `index_config.parse_preview_required` 时有意义；确认后各文档继续切分入库。 '
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmDocumentsRequest'
      responses:
        '200':
          description: 已受理
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResult'
        '400':
          $ref: '#/components/responses/InvalidParam'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/kb/{kbId}/governance:
    parameters:
    - $ref: '#/components/parameters/KbId'
    put:
      tags:
      - governance
      summary: 更新 KB 级治理开关（review_required）
      description: 'M11-CONTRACTS.md §2.2：仅影响开关翻转后新上传的文档（含新版本不重置
        存量状态），存量文档发布状态不变。'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [review_required]
              properties:
                review_required:
                  type: boolean
      responses:
        '200':
          description: 更新后的知识库
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeBaseResult'
        '400':
          $ref: '#/components/responses/InvalidParam'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/kb/{kbId}/chat-imports:
    parameters:
    - $ref: '#/components/parameters/KbId'
    post:
      tags:
      - chat-import
      summary: 聊天记录导入-匹配预览（不落库）
      description: 'M3-CONTRACTS.md §3.5：multipart 上传聊天记录导出文件，服务端调用 kb-rag-parser
        `/api/v1/parse/chat` 端点解析后，与库内既有聊天文档做会话匹配， 仅返回匹配预览，不执行导入。逻辑文档标识 = 来源渠道 + session_id：已存在的会话
        `action=NEW_VERSION`，新会话 `action=CREATE`。上传文件与解析结果暂存 MinIO， 返回的 `upload_token`
        有效期 30 分钟，供 confirm 端点使用。 M8 起格式范围扩 TXT/HTML（M8-CONTRACTS.md §0.1/§0.2），映射档案改从
        t_kb_source_mapping 解析、profile_yaml 全文随请求传给 parser；文件扩展名与所选档案 source_type
        不匹配时在预览即拦截（INVALID_PARAM）。 '
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
              - file
              properties:
                file:
                  type: string
                  format: binary
                mapping_profile:
                  type: string
                  description: 'M8 起为映射档案业务 id（t_kb_source_mapping.mapping_id）或内置名
                    （memotrace/liuhen_txt/liuhen_html，兼容 M3 旧值）；缺省按文件扩展名选默认档案
                    （M8-CONTRACTS.md §0.7/§5）'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatImportPreviewResult'
        '400':
          $ref: '#/components/responses/InvalidParam'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/kb/{kbId}/chat-imports/confirm:
    parameters:
    - $ref: '#/components/parameters/KbId'
    post:
      tags:
      - chat-import
      summary: 执行聊天记录导入
      description: 'M3-CONTRACTS.md §3.5：`upload_token` 取自匹配预览接口返回值（30 分钟内 有效），可选
        `session_ids` 只导入子集会话（缺省=预览返回的全部会话）。按会话聚合 （chat_aggregation:{window_minutes,max_messages,window_overlap}，overlap
        默认 0=顺切，M8 起支持重叠滑窗）生成 chunk_type=chat_log 的分片，metadata 写
        session_id/session_name/sender/msg_time，M8 增 window_seq/msg_span；聊天文本默认开启脱敏。 '
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatImportConfirmRequest'
      responses:
        '200':
          description: 已受理，异步执行导入
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResult'
        '400':
          $ref: '#/components/responses/InvalidParam'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: '自定义 Header：Authorization: Bearer <token>。服务端 token 表/内存缓存维护，

        有效期 24h；除 /api/v1/auth/login、/api/v1/auth/sso-available、/api/v1/auth/sso/**（含

        oidc/saml/cas 的 login 与回调）与 /actuator/** 外全部拦截。

        M15 起认证之后还有一道功能权限拦截：命中 @RequiresPermission 且主体不持有其中任一权限码时返回 403 FORBIDDEN。

        '
    ApiKeyBearer:
      type: http
      scheme: bearer
      bearerFormat: kb-sk-*
      description: '对外开放 API（/api/v1/knowledge/*）专用鉴权，Header：Authorization: Bearer

        kb-sk-*；按 key_hash（SHA-256）查验，与管理台 bearerAuth 会话令牌相互独立、走独立过滤器链，

        独立限流（按 Key 令牌桶）（M4c-CONTRACTS.md §3）。

        '
  parameters:
    KbId:
      name: kbId
      in: path
      required: true
      schema:
        type: string
      description: 知识库业务主键（kb_id）
...
  responses:
    InvalidParam:
      description: 参数校验失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorCode:
      type: string
      description: M1 错误码子集 + M4c 对外开放 API 新增错误码 + M15 FORBIDDEN，HTTP 状态码见括注
      enum:
      - INVALID_PARAM
      - UNAUTHORIZED
      - FORBIDDEN
      - NOT_FOUND
      - INVALID_API_KEY
      - API_KEY_DISABLED
      - APP_ACCESS_DENIED
      - APP_NOT_FOUND
      - VERSION_NOT_FOUND
      - VERSION_NOT_PUBLISHED
      - RATE_LIMITED
      - PARSE_FAILED
      - UPSTREAM_MODEL_ERROR
      - INTERNAL_ERROR
    ErrorResponse:
      type: object
      description: 统一错误响应
      required:
      - code
      - message
      - request_id
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          example: 'invalid parameter: name is required'
        request_id:
          type: string
          description: 全链路透传的请求 ID，经 X-Request-Id Header 贯穿 kb-server/kb-parser/模型
            Provider 日志
          example: req-3f6a2e1c
    ResultEnvelope:
      type: object
      description: 统一成功响应外壳，具体业务负载见各接口 data 字段的组合 schema
      required:
      - code
      - message
      - request_id
      properties:
        code:
          type: string
          enum:
          - OK
        message:
          type: string
          example: success
        request_id:
          type: string
    EmptyResult:
      allOf:
      - $ref: '#/components/schemas/ResultEnvelope'
      - type: object
        properties:
          data:
            nullable: true
...
    CreateKbRequest:
      type: object
      required:
      - name
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
...
    KnowledgeBase:
      type: object
      properties:
        kb_id:
          type: string
        name:
          type: string
        graph_enabled:
          type: boolean
          description: M7 GraphRAG：本库是否开启实体关系抽取与图检索路（M7-CONTRACTS.md §0.3）
        review_required:
          type: boolean
          description: M11 内容治理：开启后新上传文档初始为 DRAFT，需审核通过后才参与检索；
            仅影响开关翻转后的新上传，存量文档状态不变（M11-CONTRACTS.md §2.2）
        description:
          type: string
          nullable: true
        current_config_fingerprint:
          type: string
          nullable: true
        index_config:
          allOf:
          - $ref: '#/components/schemas/IndexConfig'
          nullable: true
          description: '索引配置（分片参数 + 父子分片，M2 新增可编辑项）；为空表示沿用 M1 固定

            默认值（chunk_max_tokens=600 / chunk_overlap=100，父子分片关闭）。

            '
        created_at:
          type: string
          format: date-time
    KnowledgeBaseResult:
      allOf:
      - $ref: '#/components/schemas/ResultEnvelope'
      - type: object
        properties:
          data:
            $ref: '#/components/schemas/KnowledgeBase'
    KnowledgeBaseListResult:
      allOf:
      - $ref: '#/components/schemas/ResultEnvelope'
      - type: object
        properties:
          data:
            type: array
            items:
              $ref: '#/components/schemas/KnowledgeBase'
...
    MetadataRule:
      type: object
      description: 'M14-CONTRACTS.md §3.1：一条配置化元数据抽取规则。key 须符合 ^[a-z][a-z0-9_]{1,31}$

        且不得与保留键冲突；type=constant 固定值（value，截 256）；type=regex 取第一个捕获组/无

        捕获组取整体（pattern ≤64，编译失败 INVALID_PARAM）；type=keyword_match 取命中子集

        （keywords ≤50 词/每词 ≤32）。'
      required:
      - key
      - type
      properties:
        key:
          type: string
          pattern: '^[a-z][a-z0-9_]{1,31}$'
        type:
          type: string
          enum:
          - constant
          - regex
          - keyword_match
        value:
          type: string
          nullable: true
          maxLength: 256
          description: constant 规则的固定值
        pattern:
          type: string
          nullable: true
          maxLength: 64
          description: regex 规则的 Java 正则源
        keywords:
          type: array
          nullable: true
          maxItems: 50
          items:
            type: string
            maxLength: 32
          description: keyword_match 规则的词表
...
    ParentChildConfig:
      type: object
      description: M2-CONTRACTS.md §1.4：父子分片，引擎只索引子片，父片正文只在 MySQL
      properties:
        enabled:
          type: boolean
          default: false
        parent_max_tokens:
          type: integer
          default: 1200
          minimum: 1
        child_max_tokens:
          type: integer
          default: 400
          minimum: 1
        child_overlap:
          type: integer
          default: 50
          minimum: 0
    SplitStrategy:
      type: string
      description: '切分策略。fixed_length=按 token 长度顺切（M1 默认）；llm_semantic=对话模型只输出

        切割点、结果落库缓存，**要求已配置对话模型**；separator=按字面量/正则分隔符切段后贪心装包

        （M14，配 split_separator/split_separator_is_regex）；heading=按 markdown 标题层级开新段

        （M14，配 split_heading_level，全文无标题回落 fixed_length）；page=按解析产物分页边界切分

        （M14，无 pages 的格式整文一页）。separator/heading/page 与 parent_child 互斥（INVALID_PARAM）；

        超长段统一回落 fixed_length 二次切分。未知取值一律 INVALID_PARAM（校验先于写入）。'
      enum:
      - fixed_length
      - llm_semantic
      - separator
      - heading
      - page
      default: fixed_length
    IndexConfig:
      type: object
      description: '知识库索引配置（t_kb_knowledge_base.index_config JSON）。chunk_max_tokens/

        chunk_overlap 为 M1 既有按长度切分参数（默认 600/100，见 M1-CONTRACTS.md

        索引管线说明），M2 起可通过 `PUT /api/v1/kb/{kbId}/index-config` 编辑；

        parent_child 为 M2 新增二级切分开关；clean_rules/parse_preview_required/

        chat_aggregation 为 M3 新增（docs/M3-CONTRACTS.md §3.3/§3.4/§3.5）。

        '
      properties:
        split_strategy:
          $ref: '#/components/schemas/SplitStrategy'
        split_separator:
          type: string
          nullable: true
          minLength: 1
          maxLength: 64
          description: 'M14 separator 策略分隔符（1-64 字符，默认 \n\n）；split_separator_is_regex=true

            时作为 Java 正则，编译失败 INVALID_PARAM；其余策略忽略此字段。'
        split_separator_is_regex:
          type: boolean
          default: false
          description: split_separator 是否为正则（M14 separator 策略）。
        split_heading_level:
          type: integer
          minimum: 1
          maximum: 6
          default: 2
          description: 'M14 heading 策略的 markdown 标题层级（1-6，默认 2）；其余策略忽略此字段。'
        metadata_rules:
          type: array
          nullable: true
          maxItems: 10
          description: 'M14-CONTRACTS.md §3.1：KB 级配置化元数据抽取规则（≤3 类共 ≤10 条），切分后逐

            chunk 抽取入 metadata 并镜像引擎（ext_ 前缀）；计入 chunk 指纹，改规则 config_stale

            重建后生效。'
          items:
            $ref: '#/components/schemas/MetadataRule'
        multimodal_enabled:
          type: boolean
          default: false
          description: 'M14-CONTRACTS.md §6.2：KB 级多模态整页索引开关，默认关。开启且多模态 provider 已

            配置时，IMAGE chunk（含扫描页整页渲染）额外产多模态向量，检索新增第三召回路；计入指纹。'
        embedding_model:
          type: string
          nullable: true
          description: 当前索引所用嵌入模型，服务端赋值；PUT 请求体无此字段，不可回写
        chunk_max_tokens:
          type: integer
          default: 600
          minimum: 1
          description: 定长切分的分段长度；LLM 语义切分下作为模型输出非法/超长时的降级兜底上限
        chunk_overlap:
          type: integer
          default: 100
          minimum: 0
        parent_child:
          $ref: '#/components/schemas/ParentChildConfig'
        clean_rules:
          $ref: '#/components/schemas/CleanRules'
        parse_preview_required:
          type: boolean
          default: false
          description: 'M3-CONTRACTS.md §3.4：默认 false，保证批量上传顺畅。开启后管线在清洗完成后 暂停，文档 process_status=PENDING_CONFIRM，需调用
            confirm 端点才继续切分入库。 '
        chat_aggregation:
          $ref: '#/components/schemas/ChatAggregationConfig'
        hide_parent_with_disabled_child:
          type: boolean
          default: false
          description: when true a parent chunk holding any disabled child is not
            returned at all
        inherit_disable_annotation:
          type: boolean
          default: true
          description: carry disable annotations to a new version when the chunk text
            hash is identical
    CleanRules:
      type: object
      description: 'M3-CONTRACTS.md §3.3：知识库级清洗规则，索引管线 clean 阶段按固定顺序执行—— 去页眉页脚 → 去水印
        → 正则替换 → 脱敏，每步可独立开关；参与 parse_fingerprint 计算 （六元组之一，见需求文档 §4.1），改清洗规则会正确标记文档
        config_stale。 '
      properties:
        strip_header_footer:
          type: boolean
          default: false
          description: 跨页重复行检测（同一文本在 >=60% 页面的首/末两行出现即视为页眉页脚并删除）
        strip_watermark_patterns:
          type: array
          default: []
          items:
            type: string
          description: 水印正则列表，匹配到的文本整体删除
        regex_replacements:
          type: array
          default: []
          items:
            $ref: '#/components/schemas/RegexReplacement'
        excel_header_join:
          type: boolean
          default: true
          description: Excel/CSV 行级切分时是否把表头拼接进每一行
        extract_metadata:
          type: boolean
          default: false
        desensitize:
          $ref: '#/components/schemas/DesensitizeConfig'
    RegexReplacement:
      type: object
      properties:
        pattern:
          type: string
        replacement:
          type: string
    DesensitizeConfig:
      type: object
      description: '正则掩码：手机号保留前 3 后 4、身份证保留前 6 后 4、银行卡保留后 4、邮箱保留域名。 聊天记录导入时默认开启（需求文档
        §4.2）。 '
      properties:
        enabled:
          type: boolean
          default: false
        phone:
          type: boolean
          default: true
        id_card:
          type: boolean
          default: true
        bank_card:
          type: boolean
          default: true
        email:
          type: boolean
          default: false
    ChatAggregationConfig:
      type: object
      description: 'M3-CONTRACTS.md §3.5：聊天记录按“会话 + 时间窗口”聚合为 chunk_type=chat_log 的分片，一期无重叠顺切（步长=窗口）；M8-CONTRACTS.md
        §0.5 增 window_overlap 重叠滑窗（默认 0 保持顺切）。 '
      properties:
        window_minutes:
          type: integer
          default: 60
          minimum: 1
        max_messages:
          type: integer
          default: 50
          minimum: 1
        window_overlap:
          type: integer
          default: 0
          minimum: 0
          description: 'M8 新增：下一窗口重复上一窗口末尾的消息数，0=既有顺切；约束
            window_overlap×2 < max_messages（每条消息至多落两个窗口），违规 INVALID_PARAM。
            开启后聊天 chunk metadata 携带 window_seq 与 msg_span（M8-CONTRACTS.md §0.5）'
    RebuildRequest:
      type: object
      description: M2-CONTRACTS.md §4：doc_ids 缺省 = 该知识库下全部 config_stale=1 文档
      properties:
        doc_ids:
          type: array
          nullable: true
          items:
            type: string
    ConfirmDocumentsRequest:
      type: object
      description: M3-CONTRACTS.md §3.4：doc_ids 缺省 = 该知识库下全部 PENDING_CONFIRM 状态文档
      properties:
        doc_ids:
          type: array
          nullable: true
          items:
            type: string
...
    ChatImportAction:
      type: string
      description: 'M3-CONTRACTS.md §3.5：逻辑文档标识 = 来源渠道 + session_id。CREATE=库内无匹配，
        建新文档；NEW_VERSION=已存在同一 session_id 的聊天文档，生成新版本（全量替换，走 M1 版本机制）。 '
      enum:
      - CREATE
      - NEW_VERSION
    TimeRange:
      type: object
      description: epoch 毫秒时间范围（该会话内消息 send_time 的最小/最大值）
      properties:
        from:
          type: integer
          format: int64
          nullable: true
        to:
          type: integer
          format: int64
          nullable: true
    ChatImportSession:
      type: object
      description: M3-CONTRACTS.md §3.5：聊天记录导入匹配预览的单个会话条目
      properties:
        session_id:
          type: string
        session_name:
          type: string
        message_count:
          type: integer
        time_range:
          $ref: '#/components/schemas/TimeRange'
        matched_doc_id:
          type: string
          nullable: true
          description: 库内已存在同一 session_id 的聊天文档 doc_id；无匹配为 null
        action:
          $ref: '#/components/schemas/ChatImportAction'
    ChatImportPreviewData:
      type: object
      properties:
        upload_token:
          type: string
          description: 上传文件与解析结果在 MinIO 的暂存凭证，有效期 30 分钟，供 confirm 端点使用
        sessions:
          type: array
          items:
            $ref: '#/components/schemas/ChatImportSession'
        skipped:
          type: object
          additionalProperties:
            type: integer
          description: 解析阶段被丢弃的消息数，按原因分组（如语音/视频等无文本行）；
            无丢弃时该键不返回
    ChatImportPreviewResult:
      allOf:
      - $ref: '#/components/schemas/ResultEnvelope'
      - type: object
        properties:
          data:
            $ref: '#/components/schemas/ChatImportPreviewData'
    ChatImportConfirmRequest:
      type: object
      required:
      - upload_token
      properties:
        upload_token:
          type: string
        session_ids:
          type: array
          nullable: true
          description: 可选子集，缺省 = 匹配预览返回的全部会话
          items:
            type: string
...